added ability to fast cache 404's, and have max-age set to 0 or intelligent settings

+++ src/Servebolt/Utils/EnvFile/Reader.php
@@ -148,7 +148,12 @@ class Reader
         }
 
         // When not it testing mode. Automatically clear the filepath if it gets this far.
-        if( !isTesting() && isset($_SERVER['DOCUMENT_ROOT']) && strpos($_SERVER['DOCUMENT_ROOT'], dirname($pathFromCache) ) === false) {
+        if( !isTesting() &&
+            empty(dirname($pathFromCache)) == false &&
+            isset($_SERVER['DOCUMENT_ROOT']) &&
+            strpos($_SERVER['DOCUMENT_ROOT'], dirname($pathFromCache) ) === false
+            )
+        {
             error_log('[Servebolt Optimizer] Is automatically clearing the environment filepath to force an update.');
             deleteOption($this->optionsKey);
         }
